Moving classes not related to client to proxy module
