+++ components/proxy/src/main/java/com/hotels/styx/applications/ApplicationsProvider.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.client.applications;
+package com.hotels.styx.applications;
 
 
 import java.util.function.Supplier;
