+++ components/proxy/src/test/java/com/hotels/styx/admin/dashboard/DashboardDataTest.java
@@ -27,7 +27,7 @@ import com.hotels.styx.api.metrics.MetricRegistry;
 import com.hotels.styx.api.metrics.codahale.CodaHaleMetricRegistry;
 import com.hotels.styx.client.StyxHostHttpClient;
 import com.hotels.styx.api.service.BackendService;
-import com.hotels.styx.client.applications.BackendServices;
+import com.hotels.styx.applications.BackendServices;
 import com.hotels.styx.infrastructure.MemoryBackedRegistry;
 import com.hotels.styx.api.service.spi.Registry;
 import org.testng.annotations.BeforeMethod;
@@ -40,7 +40,7 @@ import static com.hotels.styx.api.Id.id;
 import static com.hotels.styx.api.client.Origin.newOriginBuilder;
 import static com.hotels.styx.api.client.RemoteHost.remoteHost;
 import static com.hotels.styx.api.service.BackendService.newBackendServiceBuilder;
-import static com.hotels.styx.client.applications.BackendServices.newBackendServices;
+import static com.hotels.styx.applications.BackendServices.newBackendServices;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singleton;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
