+++ components/proxy/src/test/java/com/hotels/styx/applications/BackendServicesTest.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.client.applications;
+package com.hotels.styx.applications;
 
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.service.BackendService;
