+++ components/proxy/src/test/java/com/hotels/styx/applications/yaml/YamlApplicationsProviderTest.java
@@ -19,8 +19,8 @@ import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.service.BackendService;
 import com.hotels.styx.api.service.RewriteConfig;
 import com.hotels.styx.api.service.TlsSettings;
-import com.hotels.styx.client.applications.ApplicationsProvider;
-import com.hotels.styx.client.applications.BackendServices;
+import com.hotels.styx.applications.ApplicationsProvider;
+import com.hotels.styx.applications.BackendServices;
 import org.hamcrest.CoreMatchers;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
